feat: 디렉토리 구조 재설계 및 모듈 분리 작업 - 서비스, 페이지, 라우터, 컴포넌트 디렉토리 구조를 재설계하여 가독성 향상 - 각 기능별로 파일을 분리하여 유지보수성 개선
- 어려웠던 점: Router와 UserPreferences, 다른 컴포넌트 간의 연동 문제 발생
- 원인: 컴포넌트 간 상태 관리 및 데이터 흐름이 명확하지 않음
- 해결방법: 디버깅을 통해 문제를 파악하고, LocalStorage와 상태 관리 연동 로직을 수정하여 정상 작동 구현
